backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.0
pydantic==2.10.5
pydantic-settings==2.7.1
openai==1.59.4
python-dotenv==1.0.1
requests==2.32.3
python-multipart==0.0.20
httpx==0.28.1

pytest==7.4.0
pytest-cov==4.1.0
pytest-asyncio==0.21.0
httpx==0.25.0
black==23.12.0
flake8==6.1.0
mypy==1.7.0